+++ src/cli/Option.cpp
@@ -1,7 +1,9 @@
 
 #include "pbcopper/cli/Option.h"
+#include <stdexcept>
 using namespace PacBio;
 using namespace PacBio::CLI;
+using namespace PacBio::JSON;
 using namespace std;
 
 namespace PacBio {
@@ -48,3 +50,38 @@ const Option& Option::DefaultVerboseOption(void)
 
 const Option& Option::DefaultVersionOption(void)
 { return internal::defaultVersionOption; }
+
+std::string Option::TypeId(void) const
+{
+    const bool hasChoices = HasChoices();
+    const auto& type = DefaultValue().type();
+    switch (type)
+    {
+        case Json::value_t::number_integer  : // fall through
+        case Json::value_t::number_unsigned :
+        {
+            return (hasChoices ? "choice_integer" : "integer");
+        }
+
+        case Json::value_t::number_float :
+        {
+            return (hasChoices ? "choice_float" : "float");
+        }
+        case Json::value_t::string :
+        {
+            return (hasChoices ? "choice_string" : "string");
+        }
+        case Json::value_t::boolean :
+        {
+            return "boolean";
+        }
+
+        // unsupported/invalid JSON types
+        case Json::value_t::array     : // fall through
+        case Json::value_t::null      : // .
+        case Json::value_t::object    : // .
+        case Json::value_t::discarded : // .
+        default:
+            throw std::runtime_error{ "PacBio::CLI::ToolContract::JsonPrinter - unknown type for option: "+ Id() };
+    }
+}